ADuC7023QSPZ1: Updated Timer example

+++ Device/ADI/ADuC7023/Driver/GPIO_ADUC7023.c
@@ -85,7 +85,7 @@ void GPIO_PinConfig(GPIO_PIN_t pin, const GPIO_PIN_CFG_t *cfg, const GPIO_RESOUR
   dat = *gpio->DAT & ~(1UL << dat_pos);
 
   par_cur = *gpio->PAR;
-  par_new = (par_cur & ~(7UL << pin_pos)) | (((uint32_t)(cfg->pull_mode | cfg->strength) & 7UL) << pin_pos);
+  par_new = (par_cur & ~(7UL << pin_pos)) | ((((uint32_t)cfg->pull_mode | (uint32_t)cfg->strength) & 7UL) << pin_pos);
   if (par_cur != par_new) {
     *gpio->PAR = par_new;
   }
